Fix Visual Studio build (/wX)
warning C4800: 'uint32_t': forcing value to bool 'true' or 'false' (performance warning)
warning C4091: 'typedef ': ignored on left of '' when no variable is declared
warning C4996: 'chdir': The POSIX name for this item is deprecated.
etc

+++ src/tools/ToolsWotLK/vmap_tools/vmap4_extractor/adtfile.cpp
@@ -199,7 +199,7 @@ bool ADTFile::init(uint32 map_num, uint32 tileX, uint32 tileY)
             }
         }
         //======================
-        ADT.seek(nextpos);
+        ADT.seek(static_cast<int>(nextpos));
     }
     ADT.close();
     fclose(dirfile);
